[REFACTOR] 프로필에서 비밀번호 변경 페이지로 라우팅 시 모달 비활성화
